Rename middleware -> plugins Add api proxy middleware
